# ChatApp/API/MultiPlat-chatapp-API-master/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Chat Application",
        "description": "Documentation de l'API pour l'application de chat",
        "contact": {
            "email": "support@exemple.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/chatappAPI/mfa/toggle": {
            "post": {
                "tags": [
                    "MFA"
                ],
                "summary": "Activer ou désactiver l'authentification à deux facteurs",
                "operationId": "256851f499a71d4aba4e78d3fc712d0c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mfa_enabled"
                                ],
                                "properties": {
                                    "mfa_enabled": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "password": {
                                        "description": "Requis uniquement pour désactiver le MFA",
                                        "type": "string",
                                        "example": "MyPassword123!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "MFA modifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "MFA activé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/chatappAPI/mfa/verify": {
            "post": {
                "tags": [
                    "MFA"
                ],
                "summary": "Vérifier le code MFA et obtenir le token JWT",
                "operationId": "6c2d3f96b4632bd976742cd41b1b0b67",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "code"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "temp_token": {
                                        "type": "string",
                                        "example": "temporary_jwt_token"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Code vérifié, token JWT retourné"
                    }
                }
            }
        },
        "/chatappAPI/mfa/resend": {
            "post": {
                "tags": [
                    "MFA"
                ],
                "summary": "Renvoyer le code MFA",
                "operationId": "511337f139e94745ecf403d63f85ac2b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "temp_token"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "temp_token": {
                                        "type": "string",
                                        "example": "temporary_jwt_token"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Code renvoyé avec succès"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}